+++ framework/src/main/java/network/rtcp/type/base/sdes/SdesChunk.java
@@ -25,7 +25,7 @@ public class SdesChunk {
     ////////////////////////////////////////////////////////////
     // CONSTRUCTOR
     public SdesChunk(long ssrc, List<SdesItem> sdesItemList) {
-        this.ssrc = ssrc;
+        this.ssrc = (int) ssrc;
         this.sdesItemList = sdesItemList;
     }
